requirements.txt
argon2-cffi==20.1.0
asgiref==3.3.4
async-generator==1.10
attrs==21.2.0
autopaths==1.5.0
backcall==0.2.0
beautifulsoup4==4.9.3
biopython==1.79
bleach==3.3.0
blis==0.7.4
catalogue==2.0.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
chart-studio==1.1.0
click==7.1.2
colorama==0.4.4
colorlover==0.3.0
cufflinks==0.17.3
cycler==0.10.0
cymem==2.0.5
decorator==5.0.9
defusedxml==0.7.1
Django==3.2.3
django-cors-headers==3.7.0
djangorestframework==3.12.4
entrypoints==0.3
fastai==2.3.1
fastcore==1.3.20
fastprogress==1.0.0
idna==2.10
ipykernel==5.5.5
ipython==7.24.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kaleido==0.2.1
kiwisolver==1.3.1
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
multitasking==0.0.9
murmurhash==1.0.5
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numpy==1.20.3
packaging==20.9
pandas==1.2.4
pandas-datareader==0.9.0
pandocfilters==1.4.3
parso==0.8.2
pathy==0.5.2
peppercorn==0.6
pickleshare==0.7.5
Pillow==8.2.0
plotly==4.14.3
plumbing==2.10.4
preshed==3.0.5
prometheus-client==0.11.0
prompt-toolkit==3.0.18
psutil==5.8.0
py==1.10.0
pycparser==2.20
pydantic==1.7.4
Pygments==2.9.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
pywin32==301       --error on aws
pywinpty==1.1.1    --error om aws
PyYAML==5.4.1
pyzmq==22.1.0
requests==2.25.1
retry==0.9.2
retrying==1.3.3
scikit-learn==0.24.2
scipy==1.6.3
Send2Trash==1.5.0
sh==1.14.2
six==1.16.0
smart-open==3.0.0
soupsieve==2.2.1
spacy==3.0.6
spacy-legacy==3.0.5
sqlparse==0.4.1
srsly==2.4.1
ta==0.7.0
terminado==0.10.0
testpath==0.5.0
thinc==8.0.3
threadpoolctl==2.1.0
torch==1.8.1
torch-nightly==1.0.0.dev20190325
torchtext==0.9.1
torchvision==0.9.1
tornado==6.1
tqdm==4.61.0
traitlets==5.0.5
typer==0.3.2
typing-extensions==3.10.0.0
urllib3==1.26.4
wasabi==0.8.2
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
yfinance==0.1.59
